feat: add time period manipulation (#147)
This commit introduces the `agg.Interval` interface, and implements it
for `agg.IKEAWeek`, `agg.YearMonth` and `agg.Year`. The `agg.Interval`
interface allows for adding and subtracting periods, and determining the
start date and the end date of an interval.

This provides a clean way to, for example, find out what is the last day
of an IKEA week, what was the start date of the IKEA week that happened
6 weeks before the current one, or what are the start and end dates for
  the next month.

I am using this in the storage capacity service to enable determining
the starting point of storage projections.