feat: add otel middleware to fast echo (#206)
## Which problem is this PR solving?
Adding otel middleware to fast echo. The pr also includes fixes for the
otel config in the fast echo. Additionaly there are two span attributes
added to be able to correlate traces with logs and metrics view. The
link is working with span attributes that is why the service name and
the environment are added to attributes
https://jira.digital.ingka.com/browse/OP-3911