feat(fastecho): Allow server customization (#208)
We sometimes need to have more control over the server than just running
it until it interrupts. This feature allows for the FastEcho server to
be initialized separately from actually starting it by using the new
methods that give a `FastEcho` instance. In this way we can control when
the server starts, integrate it with another HTTP server, and control
how it stops. This is particularly useful in testing.