feat: enable customising underlying Echo object (#54)
# One-line summary

Adds support for accessing the underlying Echo object.

## Description
Currently, the only way to access the Echo object is via the Plugin
system. So, if we want to, say add a service level middleware, we'd have
to create a plugin and bind the middleware via the echo instance in
Plugin interface.

With this, we can access and manipulate the Echo object directly by
passing an `EchoFn` which accepts the underlying echo object as
argument.

## Types of Changes
_What types of changes does your code introduce? Keep the ones that
apply:_

- Refactor/improvements
- Documentation / non-code

---------

Co-authored-by: Adrian Ciobanu <97035640+AdrianCio@users.noreply.github.com>